refactor(plugin): Rename and refactor Tauri Plugin
This commit renames the plugin to tauri-plugin-devtools based on Lucas's suggestion. Adopting the devtools namespace is strategic to ensure it's reserved for this purpose.

The plugin integrates the functionalities of both the `inspector_protocol_server` and the `inspector_protocol_subscriber`. This design allows flexibility, indicating that the underlying implementation can be swapped out or repurposed for different scenarios like `EMP`.

Several enhancements have been made to the documentation, aiming for clarity and completeness. There's undoubtedly room for further improvement, but this serves as a starting point.

Significant changes include:

- Direct implementation of the tauri::plugin::Plugin trait in Devtools, eliminating reliance on TauriPlugin.
- Removal of the internal channel previously capturing AppReady and other event loop events. Thanks to the new Devtools construction, direct updates to the Metrics are now possible.

Signed-off-by: david-crabnebula <david@crabnebula.dev>