feat: track thread ID
Actually track the ID of the thread entering and exiting a span, instead of sending a zero value.

The thread ID does **not** correspond to the OS'es underlying thread ID, but it is guaranteed to be a unique identifier of the thread.

It is also not guaranteed that the thread ID will return the same values in future versions.