feat(sources): use actual app aseets on production app
This commit changes the Sources backend implementation to use Tauri's asset_resolver (only available on production or when devPath points to the filesystem). It doesn't make sense to use the current dir as reference on production, so this aligns better with what a user expects when instrumenting a debug built application (tauri build --debug) instead of tauri dev.