chore(example): use a dev server
This changes our examples to look like an actual Tauri app on dev mode: uses a dev server and does not embed any files in the binary. This will change the devtools asset resolver to actually look for files on the current working directory, which helps the Config tab for now. In the future we might tweak the getEntryBytes a little to look for Tauri config files in parent directories so it works better for built apps (apps in the target/debug folder).