feat: add `tauri-plugin-log` compat
Since devtools is incompatible with the log plugin anyway (it serves the same purpose too but better) it would be handy for it to fully replace the log plugin during development.

This PR adds the last missing piece of interop with the log plugin by intercepting the JS log messages and emitting them to the same log stream as all other messages.